TITLE: wazuh-agent - use deploy mod of agent, host IP from env
(dockerfile)wazuh-agent: if provided wazuh-manager ip in the ENV, the installation of agent will be in deploy mode. The registeration can also be done by bash script, which is put in config/api-register-agent.sh

(config/api-register-agent.sh)automatic registeration script: currently it is not used in anywhere. there is a to-do that adding args passing code to decord the hardcorded variables in the beginning of this script.