TITLE: wazuh-agent - accidentlySolvedNoSuchFile
by adding VOLUME, and also, at the k8s side add pv, pvc, mount the path to //entrypoint-scripts, it accidently solves the problem that nothing is installed on the COS HOST.

now I guess, the problem happened because 1. I mounted a path overriding the path of Wazuh-agent; 2. the run of docker script desires the allocation of storage, while COS strictly limit the filesystem permisson on certain paths.

and also I deduct, none of the script, the sh files, of this dockerfile is successfully run. so the second guess is not true.

and also a problem, passing Public IP of loadbalancer of wazuh-manager as the installation parameter of wazuh-agent does not make it register sucessfully.